LayoutTests/http/tests/security/postMessage/invalid-origin-throws-exception.html

<script>
if (window.layoutTestController)
  layoutTestController.dumpAsText();
</script>

<!DOCTYPE html>
<html>
<head>
<script>
if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.waitUntilDone();
}

function recv(e) {
    document.getElementById("result").innerHTML += "<br>" + e.data;
}
 
document.addEventListener("message", recv, false);

function tryPostMessage(origin) {
  try {
    win.postMessage("Trying origin=" + origin, origin);
    document.getElementById("result").innerHTML += "<br>Posted message to '" + origin + "' without any exceptions.";
  } catch(ex) {
    document.getElementById("result").innerHTML += "<br>Encountered exception " + ex + " while posting message to '" + origin + "'.";
  }
}

function test() {
    var iframe = document.getElementById('child');
    win = iframe.contentWindow;

    // Non-URLs should fail with a syntax error.
    tryPostMessage("");
    tryPostMessage("asdf");

    // URLs without an origin should fail without generating any errors.
    tryPostMessage("asdf:");
    tryPostMessage("http:");

    // WebKit converts URLs that start with / to file:// URLs. They don't match the target frame, so they fail silently.
    tryPostMessage("/tmp/foo");
    tryPostMessage("//localhost");

    if (window.layoutTestController)
        layoutTestController.notifyDone();
}
</script>
<body>
<div>window.location.href = <script>document.write(window.location.href);</script></div>
<div><iframe src="http://localhost:8000/security/postMessage/resources/post-message-listener.html" onload="test()"
  id="child" width="800" height="300" style="border: 1px solid black;">
</iframe></div>
<div id="result">waiting...</div>
</body>
</html>
