LayoutTests/http/tests/security/postMessage/invalid-origin-throws-exception-expected.txt

CONSOLE MESSAGE: line 0: Unable to post message to asdf:. Recipient has origin http://localhost:8000.

CONSOLE MESSAGE: line 0: Unable to post message to http:. Recipient has origin http://localhost:8000.

CONSOLE MESSAGE: line 0: Unable to post message to /tmp/foo. Recipient has origin http://localhost:8000.

CONSOLE MESSAGE: line 0: Unable to post message to //localhost. Recipient has origin http://localhost:8000.

window.location.href = http://127.0.0.1:8000/security/postMessage/invalid-origin-throws-exception.html

waiting...
Encountered exception Error: SYNTAX_ERR: DOM Exception 12 while posting message to ''.
Encountered exception Error: SYNTAX_ERR: DOM Exception 12 while posting message to 'asdf'.
Posted message to 'asdf:' without any exceptions.
Posted message to 'http:' without any exceptions.
Posted message to '/tmp/foo' without any exceptions.
Posted message to '//localhost' without any exceptions.
